requirements.txt
# These requirements are for running tests and the sample_site, the library 
# itself only needs a subset, see setup.py for that list
coverage==3.7.1
Django==1.8.4
django-awl==0.9.0
libsass==0.8.3
#mock==1.3.0
pudb==2015.3
pyflakes==0.9.2
Pygments==2.0.2
Sphinx==1.3.1
tox==2.1.1
twine==1.5.0
urwid==1.3.0
wheel==0.24.0
wrench==0.5.1
